+++ app/src/main/java/com/mocare/app/data/local/entity/FuelCheckpointEntity.kt
@@ -15,5 +15,5 @@ data class FuelCheckpointEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val timestamp: Long = System.currentTimeMillis(),
-    val odometerKm: Int
+    val odometerKm: Double
 )
